+++ rust/src/server/mod.rs
@@ -21,6 +21,7 @@ pub mod roots;
 use roots::has_project_marker;
 pub mod tool_trait;
 pub mod tool_visibility;
+pub mod tools_config_watch;
 
 use futures::FutureExt;
 use rmcp::ErrorData;